fastapibackend/requirements.txt
duckduckgo_search==7.1.1
phidata==2.7.6
openai==1.57.2
google-generativeai==0.7.2
fastapi==0.111.1
pydantic==2.7.4
uvicorn==0.30.3
gunicorn==23.0.0 